backend/requirements.txt
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
fastapi
sentence-transformers
uvicorn
chromadb
langchain-googledrive
langchain-community 
langchain-google-genai
langchain-huggingface
unstructured 
pandas
openpyxl
SQLAlchemy
passlib
faiss-cpu
rank_bm25
